refactor: consolidate project edit forms and remove deprecated sidebar component
